rename ts source files
