+++ src/index.ts
@@ -1,7 +1,7 @@
-import { createContract, createContractFromTx } from "./create-contract";
-import { getContract } from "./get-contract";
-import { interactWrite, interactWriteDryRun } from "./interactions";
-import { replayToState } from "./replay-contract";
+import { createContract, createContractFromTx } from "./contract-create";
+import { getContract } from "./contract-load";
+import { interactWrite, interactWriteDryRun } from "./contract-interact";
+import { replayToState } from "./contract-replay";
 
 export {
   createContract,
